move generated interfaces to .h files
