+++ src/Packet.cc
@@ -1,7 +1,7 @@
-#include "Packet.hpp"
+#include "Packet.h"
 #include <stdint.h>
 #include "string.h"
-#include "packet/GPS_Packet.cpp"
+#include "packet/GPS_Packet.h"
 
 Writeable &operator<<(Writeable &w, Packet *p)
 {
